refactor: foldering by function and method
