+++ users/handler/authorization.go
@@ -1,0 +1,35 @@
+package handler_users
+
+import (
+	"net/http"
+	err_conv "ppob/helper/err"
+	"ppob/users/handler/request"
+
+	"github.com/go-playground/validator/v10"
+	"github.com/labstack/echo/v4"
+)
+
+func (uh *UsersHandler) Authorization(ctx echo.Context) error {
+	req := request.RequestJSONLogin{}
+	ctx.Bind(&req)
+	if err := uh.Validation.Struct(req); err != nil {
+		stringerr := []string{}
+		for _, errval := range err.(validator.ValidationErrors) {
+			stringerr = append(stringerr, errval.Field()+" is not "+errval.Tag())
+		}
+		return ctx.JSON(http.StatusBadRequest, stringerr)
+	}
+
+	res, err := uh.usecase.Login(req.Email, req.Password)
+	if err != nil {
+		return err_conv.Conversion(err, ctx)
+	}
+
+	return ctx.JSON(http.StatusOK, map[string]interface{}{
+		"message": "user success login",
+		"rescode": http.StatusOK,
+		"data": map[string]interface{}{
+			"token": res,
+		},
+	})
+}
